% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWorldClim.R
\name{getWorldClim}
\alias{getWorldClim}
\title{getWorldClim}
\usage{
getWorldClim(res, country, path)
}
\arguments{
\item{res}{numeric, resolution of the data. Possible values are 10, 5, 2.5,
or 0.5 arc minutes.}

\item{country, }{character, country name.}

\item{path, }{character, the path to the folder wherer the variables shall
be saved}
}
\value{
This function downloads all 19 climatic variables and returns them
in a raster stack.
}
\description{
Downloads bioclimatic data from WorldClim
}
\examples{
world_clim <- getWorldClim(res = 0.5,country = "Czech Republic",
path = getwd())
}
