% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natAlPoints.R
\name{natAlPoints}
\alias{natAlPoints}
\title{natAlPoints}
\usage{
natAlPoints(occ, nat_al)
}
\arguments{
\item{occ}{able containing columns with the species name, longitude, and
latitude.}

\item{nat_al}{list containing reference region shapefiles for the native and
alien distribution, having all alien regions overlapping native regions
eliminated. Output of function \strong{noNatAlOverlap}}
}
\value{
list containing points overlapping the native regions, and those
overlapping the alien regions.
}
\description{
Identifies native and alien points based on overlap with regions in
checklists.
}
