% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheBreadth.R
\name{nicheBreadth}
\alias{nicheBreadth}
\title{nicheBreadth}
\usage{
nicheBreadth(occ, path, var_cont)
}
\arguments{
\item{occ}{table containing columns with the species name, longitude, and
latitude.}

\item{path}{character, the path to the folder containing the variables.}

\item{var_cont}{table informing the contribution of each variable in a
maxent model, the output of \strong{variableContribution}.}
}
\value{
This function return a table informing which variables are less
correlated than the defined threshold.
}
\description{
Calculate the n-dimensional niche breadth based on point records and
environmental layers.
}
