% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNatAlOverlap.R
\name{noNatAlOverlap}
\alias{noNatAlOverlap}
\title{noNatAlOverlap}
\usage{
noNatAlOverlap(ref_reg)
}
\arguments{
\item{ref_reg}{list containing three reference region shapefiles showing the
regions where the species is present, native, and alien. Output of the
\strong{bRacatus} function \strong{giftRegion}.}
}
\value{
list containing reference region shapefiles for the native and alien
distribution, having all alien regions overlapping native regions
eliminated.
}
\description{
Eliminate from the shapefiles regions listed as alien that overlap
regions listed as native.
}
