% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoints.R
\name{plotPoints}
\alias{plotPoints}
\title{plotPoints}
\usage{
plotPoints(
  occ,
  location = "location",
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  colours = "standard"
)
}
\arguments{
\item{occ}{table containing columns with the species name, longitude, and
latitude}

\item{lon}{character, name of the longitude column, default is
"decimalLongitude".}

\item{lat}{character, name of the latitude column, default is
"decimalLatitude".}

\item{colours}{character, options are "bicha" or "standard". Default is
"standard"}

\item{species}{character, name of the column containing species binomial
name. Default is "species".}
}
\value{
This function plots occurrences on a background.
}
\description{
Plot occurrence points with background for visualisation.
}
