% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableContribution.R
\name{variableContribution}
\alias{variableContribution}
\title{variableContribution}
\usage{
variableContribution(occ, path, corel)
}
\arguments{
\item{occ}{table containing columns with the species name, longitude, and
latitude.}

\item{path}{character, the path to the folder containing the variables.}

\item{corel}{table, output of the function \strong{variableCorrelation}.}
}
\value{
This function return a table informing the contribution of each
variable in a maxent model.
}
\description{
Tests of the environmental variable importance for a given species based on
Maxent algorithm.
}
