% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableCorrelation.R
\name{variableCorrelation}
\alias{variableCorrelation}
\title{variableCorrelation}
\usage{
variableCorrelation(occ, path, th = 0.7)
}
\arguments{
\item{occ}{table containing columns with the species name, longitude, and
latitude.}

\item{path}{character, the path to the folder containing the variables.}

\item{th}{threshold of correlation. Default is 0.7.}
}
\value{
This function return a table informing which variables are less
correlated than the defined threshold.
}
\description{
Checks the correlation amongst the variables that contributes most to the
species modelling with maxent.
}
