% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablePreparation.R
\name{variablePreparation}
\alias{variablePreparation}
\title{variablePreparation}
\usage{
variablePreparation(new_path, var_cont, th = 10, original_format = ".asc")
}
\arguments{
\item{new_path}{character, the path to the folder containing the
standardised variables.}

\item{var_cont}{table informing the contribution of each variable in a
maxent model, the output of \strong{variableContribution}.}

\item{th}{numeric, threshold of variable contribution to the model. Default
is 10\%.}

\item{original_format}{character, indicates the format of the layers.}
}
\value{
This function returns a raster stack containing the variables that
contributed to the model more than the selected threshold (in percentage).
}
\description{
Prepares the standardised versions of the variables that contribute the most
in the maxent model for the niche breadth calculation.
}
