% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableStandardisation.R
\name{variableStandardisation}
\alias{variableStandardisation}
\title{variableStandardisation}
\usage{
variableStandardisation(original_path, new_path, original_format = ".asc")
}
\arguments{
\item{original_path}{character, the path to the folder containing the
variables.}

\item{new_path}{character, the path to the folder where the standardised
variables shall be saved.}

\item{original_format}{character, indicates the format of the layers to be
standardised.}
}
\value{
This function standardises the variables and saves the new version.
}
\description{
Standardise raster layers by subtracting the mean and dividing by the
standard deviation (Cardillo et al. 2018 Journal of Biogeography).
}
