% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWorldClim.R
\name{getWorldClim}
\alias{getWorldClim}
\title{getWorldClim}
\usage{
getWorldClim(res, country = NULL, path)
}
\arguments{
\item{res}{numeric, resolution of the data. Possible values are 10, 5, 2.5,
or 0.5 arc minutes. If res = 0.5, country must be specified.}

\item{country}{character, country name. Sets the tile containing the central
coordinates of the defined country. If NULL, gets variable for whole world.
Default is NULL.}

\item{path}{character, the path to the folder where the variables shall
be saved.}
}
\value{
This function downloads all 19 climatic variables and saves them in
the folder defined in "path".
}
\description{
Downloads bioclimatic data from WorldClim
}
\examples{
world_clim <- getWorldClim(res = 0.5,country = "Czech Republic",
path = getwd())
}
