% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheBreadth.R
\name{nicheBreadth}
\alias{nicheBreadth}
\title{nicheBreadth}
\usage{
nicheBreadth(occ, vars2)
}
\arguments{
\item{occ}{table containing columns with the species name, longitude, and
latitude.}

\item{vars2}{raster stack containing the variables that contributed most to
the model. The output of \strong{variablePreparation}.}
}
\value{
This function returns the niche breadth calculated based on the
variables' values in the occ locations.
}
\description{
Calculate the n-dimensional niche breadth based on point records and
environmental layers.
}
